Move addClipMarkup into its own file
